Refactor static methods in EagerTagDecorator into separate class
